Rename Xgit.Plumbing.CommitTree.run/2 to Xgit.Repository.Plumbing.commit_tree/2.
